feat(i18n): co-locate locale files
